Add delete methods
This also removes the repository registry. Haven't that and object
mgr felt reduntant. While having individual repositories (or a
collection of known ones) passed into the class would be nice, it
just seems wrong when our object manager can get them and especially
when different repositories that are outside of the om are passed in
along side the om.

This also adds the event dispatcher